refactor : [CMA-5] java to kotlin migration
- LikeService 개선
- postItem 개선
- postResponse 추가
- Mate Dto userId type을 Long으로 변경
- GoalDateInfo 추가 (타임라인 사용)

+++ app/src/main/java/kr/co/company/capstone/dto/goal/GoalDetail.kt
@@ -1,7 +1,7 @@
 package kr.co.company.capstone.dto.goal
 
 data class GoalDetail(
-    val goalId: Int,
+    val goalId: Long,
     val category: String,
     val title: String,
     val startDate: String,
